Procedural Dungeon + modifications
Set up procedural generated dungeon by following a tutorial, I also personally set up random pick up spawns on each room (check DungeonCreator.cs), it also checks whether if that item is too close to the doorway then it shouldn't be spawned, furthermore I created box colliders on each mesh that is floor and walls, I sized and centered them aswell since at default there were only 1x1 cubes (see AddBoxCollider method in DungeonCreator.cs). Finally there is the GenerateNavMesh method which generates a navigation mesh to be able used by the enemy AI, I achieved this by modifying dungeonFloor to be a typeof NavMeshSurface and then in the method I ran BuildNavMesh() on it.

I'm planning to have the aforementioned methods in their own scripts if possible, this is to make distinctions which scripts are mine and of course to make everything more readable.
Like, for example, how I have made BossRoom.cs and DungeonCreator.cs is using it in this iteration.