Enemies w/ Patrol, Boss Room Walls
To pass a boss room wall you can either jetpack your way in or click F when you're nearby for the player's transform to move towards the boss. New script would be PlayerInteraction that does that (its called in BossRoom.cs), it also sends out static events if the player is nearby the boss room walls, if the player is then a UI appears that instructs the player what to press when nearby the wall, the listener script is on the UI's parent.
Lastly, I created EnemyGenerator.cs, called from the main DungeonCreator script. Inside of it there's a method to generate enemies, for them to instantiate the room needs to be at least 15 by 15 in size, so they wouldn't spawn in corridors and then there's a 50 percent chance of spawning, SpawnEnemy method is then called. I instantiate the enemy and a patrol instance, with its default patrol path. In RandomizePathNodes I try to tweak these path points a bit, usually it looks like the same triangle, I might have to readjust this or perhaps create methods into adding or subtracting path nodes. Also might need a better way to check if the path node is within bounds (on the nav mesh/same room).